requirements.txt
Django==4.2
psycopg2-binary==2.9.6
python-dotenv==1.0.0
djangorestframework==3.14.0
drf-yasg==1.21.6
coreapi==2.3.3
djangorestframework-simplejwt==5.2.2
django-cors-headers==4.2.0
gunicorn==21.2.0
Pillow==9.5.0
djoser==2.2.0
django-filter==23.2
coverage==7.2.7
flake8==6.1.0
black==23.7.0